requirements.txt
python-dotenv
soundfile
numpy
scipy

websockets
fastapi[standard]
beautifulsoup4
aiohttp